move files out of components
