+++ app/webpacker/react_on_rails_components/Delegates/DelegatesOfAllRegion.jsx
@@ -5,8 +5,8 @@ import DelegatesTable from './DelegatesTable';
 import useLoadedData from '../../lib/hooks/useLoadedData';
 import { apiV0Urls } from '../../lib/requests/routes.js.erb';
 import { groupTypes } from '../../lib/wca-data.js.erb';
-import Loading from '../Requests/Loading';
-import Errored from '../Requests/Errored';
+import Loading from '../../components/Requests/Loading';
+import Errored from '../../components/Requests/Errored';
 import I18n from '../../lib/i18n';
 import dateSince from '../../lib/helpers/date-since';
