+++ app/webpacker/react_on_rails_components/StaticPages/Disclaimer.js
