+++ app/webpacker/react_on_rails_components/StaticPages/Logo.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { Button } from 'semantic-ui-react';
-import I18nHTMLTranslate from '../I18nHTMLTranslate';
+import I18nHTMLTranslate from '../../components/I18nHTMLTranslate';
 import lockupPrimary from './LogoImages/1 Positive Primary/WCA Logo Lockup.svg';
 import lockupNegative from './LogoImages/2 Negative Primary/WCA Logo Lockup.svg';
 import lockupMonoBlack from './LogoImages/3 Mono Black/WCA Logo Lockup.svg';
