+++ app/webpacker/react_on_rails_components/TeamsCommitteesCouncils/GroupPage.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Button, Header, Message } from 'semantic-ui-react';
-import EmailButton from '../EmailButton';
+import EmailButton from '../../components/EmailButton';
 import { apiV0Urls, contactRecipientUrl } from '../../lib/requests/routes.js.erb';
 import I18n from '../../lib/i18n';
 import useLoadedData from '../../lib/hooks/useLoadedData';
-import Loading from '../Requests/Loading';
-import Errored from '../Requests/Errored';
+import Loading from '../../components/Requests/Loading';
+import Errored from '../../components/Requests/Errored';
 import RolesTable from './RolesTable';
 
 export default function GroupPage({ group, canViewPastRoles }) {
