+++ app/webpacker/react_on_rails_components/Translators/index.jsx
@@ -4,10 +4,10 @@ import _ from 'lodash';
 import useLoadedData from '../../lib/hooks/useLoadedData';
 import { apiV0Urls } from '../../lib/requests/routes.js.erb';
 import { groupTypes } from '../../lib/wca-data.js.erb';
-import Loading from '../Requests/Loading';
-import Errored from '../Requests/Errored';
+import Loading from '../../components/Requests/Loading';
+import Errored from '../../components/Requests/Errored';
 import I18n from '../../lib/i18n';
-import UserBadge from '../UserBadge';
+import UserBadge from '../../components/UserBadge';
 
 export default function Translators() {
   const {
